requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.0
celery>=5.3
redis>=5.0
channels>=4.0
channels-redis>=4.2
httpx>=0.27
psycopg2-binary>=2.9
python-dotenv>=1.0
Pillow>=10.0
gunicorn>=22.0
daphne>=4.1
